Decimal And Cart Path
redirect float => decimal
cart Path update services People and Admin
edit Post - if Product in card return Conflict. But post create new resurse and not edit

+++ Models/Market/Order.cs
@@ -9,7 +9,7 @@ public class Order
     
     public string UserId { get; set; }
 
-    public float TotalPrice { get; set; } = 0;
+    public decimal TotalPrice { get; set; } = 0;
     
     public bool Status { get; set; } = false;
